+++ clients/core/src/App.tsx
@@ -4,7 +4,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { KeycloakProvider } from '@/keycloak/KeycloakProvider'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import { ManagementRoot } from './management/ManagementConsole'
-import { TemplateRoutes } from './PhaseMapping/ExternalRouters/TemplateRoutes'
+import { TemplateRoutes } from './PhaseMapping/ExternalRoutes/TemplateRoutes'
 import { PhaseRouterMapping } from './PhaseMapping/PhaseRouterMapping'
 import PrivacyPage from './LegalPages/Privacy'
 import ImprintPage from './LegalPages/Imprint'
