+++ clients/core/src/PhaseMapping/ExternalSidebars/ExternalSidebar.tsx
